add placeholder for radiant controls tests
